+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `state` feature (enabled by default) for automatic context activation based on `bevy_state` states. See the module documentation for more details.
+
 ## [0.22.1] - 2026-01-16
 
 ### Fixed